@@ -10,17 +10,17 @@
   "dependencies": {
     "core-js": "^3.6.5",
     "tailwindcss": "^1.1.2",
-    "vue": "^2.6.10",
+    "vue": "^3.0.0-beta.1",
     "vue-router": "^3.5.1"
   },
   "devDependencies": {
     "@vue/cli-plugin-babel": "~4.5.12",
     "@vue/cli-plugin-eslint": "~4.5.12",
     "@vue/cli-service": "~4.5.12",
+    "@vue/compiler-sfc": "^3.0.0-beta.1",
     "babel-eslint": "^10.1.0",
     "eslint": "^6.7.2",
-    "eslint-plugin-vue": "^6.2.2",
-    "vue-template-compiler": "^2.6.10"
+    "eslint-plugin-vue": "^7.0.0-alpha.0",
   },
   "eslintConfig": {
     "root": true,
@@ -28,7 +28,7 @@
       "node": true
     },
     "extends": [
-      "plugin:vue/essential",
+      "plugin:vue/vue3-essential",
       "eslint:recommended"
     ],
     "rules": {},
