@@ -1,16 +1,11 @@
-import Vue from 'vue'
+import { createApp } from 'vue';
 import App from './App.vue'
 import VueRouter from 'vue-router';
 import { routes } from './route';
-
-Vue.use(VueRouter);
 
 const router = new VueRouter({
   routes,
   mode: 'history'
 });
 
-new Vue({
-  router,
-  render: h => h(App)
-}).$mount('#app')
+createApp(App).use(router).mount('#app')
